Inserindo paginas de confirmação e página de detalhes

+++ src/app/component/card-baralho/card-baralho.component.ts
@@ -2,10 +2,14 @@ import { Component, OnInit, Input } from '@angular/core';
 import { BaralhoDal } from 'src/app/core/dal/baralho.dal';
 import { BaralhoModel } from 'src/app/model/baralho.model';
 import { BehaviorSubject } from 'rxjs';
-import { SalvarBaralhoDialog } from 'src/app/page/dialog/salvar-baralho.dialog';
+import { SalvarBaralhoDialog } from 'src/app/page/dialog/salvar-baralho/salvar-baralho.dialog';
 import { isNullOrUndefined } from 'util';
 import { MatDialog } from '@angular/material/dialog';
 import { BaralhoProviderService } from 'src/app/provider/baralho-provider.service';
+import { ConfirmacaoDialogModel } from 'src/app/model/confirmacao-dialogo.model';
+import { ConfirmacaoDialogComponent } from 'src/app/page/dialog/confirmacao-dialog/confirmacao-dialog.component';
+import { Router } from '@angular/router';
+import { ParamUtil } from 'src/app/util/param.util';
 
 @Component({
   selector: 'app-card-baralho',
@@ -19,7 +23,8 @@ export class CardBaralhoComponent implements OnInit {
 
   constructor(
     public dialog: MatDialog,
-    public baralhoDao: BaralhoProviderService
+    public baralhoDao: BaralhoProviderService,
+    private _router: Router,
   ) { }
 
   ngOnInit() {
@@ -27,6 +32,11 @@ export class CardBaralhoComponent implements OnInit {
 
   selecionarCard() {
     this.baralhoComunicacao.next(this.baralho);
+  }
+
+  abrirDetalhes() {
+    ParamUtil.setParam('baralho', this.baralho);
+    this._router.navigate(['/detalhes']);
   }
 
   openEditarBaralho(): void {
@@ -45,9 +55,26 @@ export class CardBaralhoComponent implements OnInit {
     });
   }
 
-  deletarBaralho(): void {
-    this.baralhoDao.dao.remover(this.baralho);
-    this.baralhoComunicacao.next(null);
+  async deletarBaralho() {
+    if (<boolean>await this.confirmacaoDialogo("Confirma remover este Baralho?", "Confirmação")) {
+      this.baralhoDao.dao.remover(this.baralho);
+      this.baralhoComunicacao.next(null);
+    }
+  }
+
+  async confirmacaoDialogo(mensagem, titulo): Promise<boolean> {
+    return await new Promise<boolean>(resolve => {
+      const dialogData = new ConfirmacaoDialogModel(titulo, mensagem);
+
+      const dialogRef = this.dialog.open(ConfirmacaoDialogComponent, {
+        maxWidth: "400px",
+        data: dialogData
+      });
+
+      dialogRef.afterClosed().subscribe(dialogResult => {
+        resolve(dialogResult);
+      });
+    });
   }
 
 }
